requirements.txt
bcrypt
certifi
cffi
charset-normalizer
click
colorama
Flask
Flask-Bcrypt
Flask-Login
Flask-SQLAlchemy
Flask-WTF
greenlet
idna
itsdangerous
Jinja2
joblib
MarkupSafe
numpy
pandas
Pillow
pycparser
python-dateutil
pytz
requests
scikit-learn
scipy
six
SQLAlchemy
threadpoolctl
torch
torchvision
typing_extensions
urllib3
Werkzeug
WTForms
gunicorn
